hyperparameters polars

+++ src/hawkes/simulation.py
@@ -141,7 +141,7 @@ def hawkes_estimation(t: np.ndarray, filename: str = "hawkes_estimation.parquet"
                'AIC': round(hawkes_process.AIC, 3)}
 
     # Written parameters to parquet file
-    write_parquet(metrics, dtype=np.float32, filename=filename)
+    write_parquet(metrics, filename=filename)
     # Transformed times so that the first observation is at 0 and the last at 1
     [t_transform, interval_transform] = hawkes_process.t_trans() 
     # Predicted the Hawkes process
